content/base/test/test_bug605982.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=605982
-->
<head>
  <title>Test for Bug 605982</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=605982">Mozilla Bug 605982</a>
<p id="display"></p>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 605982 **/

var elmt = document.createElement("div");

var caught = false;
try {
  elmt.mozMatchesSelector("!!");
} catch(e) {
  ok(e.code == DOMException.SYNTAX_ERR, "Error should be SYNTAX_ERROR");
  caught = true;
}
ok(caught, "An exception should have been thrown");

</script>
</pre>
</body>
</html>
